+++ Auxiliary/string.hpp
@@ -73,7 +73,7 @@ std::string retval; { \
 	std::string make_nstring(const wchar_t* wbytes);
 	std::string make_nstring(Platform::String^ wstr);
 
-	std::string binumber(unsigned int n, size_t bitsize = 0);
+	std::string binumber(unsigned long long n, size_t bitsize = 0);
 
 	Platform::String^ string_first_line(Platform::String^ src);
 	std::list<Platform::String^> string_lines(Platform::String^ src, bool skip_empty_line = false);
